Use python profile interface to collect python trace, including built-ins
Python ppt package collects traces using python profile interface,
including built-ins. When a built-in is called, ppt tells the pin tool
to collect the function names of mkl libraries. When a built-in
returns, ppt tells the pin tool to stop collecting and return the list
of native functions that were called. A simple native extension is
used to pass the commands and data between python and pintool.